move wait secs to the module where its used
