+++ qcodes/tests/helpers/test_waitsecs.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 import pytest
 from qcodes.logger.logger import LogCapture
-from qcodes.utils.helpers import wait_secs
+from qcodes.loops import wait_secs
 
 
 def test_bad_calls():
